refactor. JpaQueryFactory 트랜잭션 동일 위해 하나의 config로 정리 1. DB를 같은 DB로 사용중이기 때문에 transaction을 하나의 모듈로 묶음
