+++ api/src/main/java/com/commerce/saleday/api/infra/database/repository/outbox/OutboxJpaRepository.java
@@ -1,7 +1,7 @@
 package com.commerce.saleday.api.infra.database.repository.outbox;
 
-import com.commerce.saleday.domain.outbox.model.OutboxMessage;
-import com.commerce.saleday.domain.outbox.model.OutboxStatus;
+import com.commerce.saleday.common.outbox.model.OutboxMessage;
+import com.commerce.saleday.common.outbox.model.OutboxStatus;
 import java.time.LocalDateTime;
 import org.springframework.data.jpa.repository.JpaRepository;
