+++ api/src/main/java/com/commerce/saleday/api/infra/transaction/outbox/OrderEventPublishHandler.java
@@ -2,8 +2,8 @@ package com.commerce.saleday.api.infra.transaction.outbox;
 
 import static org.springframework.transaction.event.TransactionPhase.AFTER_COMMIT;
 
-import com.commerce.saleday.domain.outbox.model.OutboxMessage;
-import com.commerce.saleday.domain.outbox.repository.OutboxRepository;
+import com.commerce.saleday.common.outbox.model.OutboxMessage;
+import com.commerce.saleday.common.outbox.repository.OutboxRepository;
 import com.commerce.saleday.order.domain.stock.port.ItemStockPublisherKafkaPort;
 import com.commerce.saleday.message.stock.DecreaseStockEvent;
 import lombok.RequiredArgsConstructor;
