+++ api/src/main/java/com/commerce/saleday/api/infra/transaction/outbox/OutboxRecorder.java
@@ -1,7 +1,7 @@
 package com.commerce.saleday.api.infra.transaction.outbox;
 
-import com.commerce.saleday.domain.outbox.model.OutboxMessage;
-import com.commerce.saleday.domain.outbox.repository.OutboxRepository;
+import com.commerce.saleday.common.outbox.model.OutboxMessage;
+import com.commerce.saleday.common.outbox.repository.OutboxRepository;
 import com.commerce.saleday.message.stock.DecreaseStockEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
