+++ common/src/main/java/com/commerce/saleday/common/outbox/repository/OutboxRepository.java
@@ -1,12 +1,12 @@
-package com.commerce.saleday.domain.outbox.repository;
+package com.commerce.saleday.common.outbox.repository;
 
-import com.commerce.saleday.domain.outbox.model.OutboxMessage;
+import com.commerce.saleday.common.outbox.model.OutboxMessage;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface OutboxRepository {
 
-  public OutboxMessage save(OutboxMessage outboxMessage);
+  OutboxMessage save(OutboxMessage outboxMessage);
 
   List<OutboxMessage> saveAll(List<OutboxMessage> failedMessages);
