+++ consumer/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
     implementation project(':message')      // 메세지 의존성 추가
-    implementation project(':domain')      // domain module 의존성 추가
+    implementation project(':common')      // common module 의존성 추가
     implementation project(':order')      // order module 의존성 추가
 }
