+++ consumer/src/main/java/com/commerce/consumer/ConsumerApplication.java
@@ -1,13 +1,16 @@
 package com.commerce.consumer;
 
+import com.commerce.saleday.common.jpa.config.QuerydslConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @EnableJpaAuditing
 @SpringBootApplication
-@EntityScan(basePackages = "com.commerce.saleday")//모듈로 옮겼기 떄문에 scan 달아줌
+@EntityScan(basePackages = "com.commerce.saleday")
+@Import(QuerydslConfig.class)
 public class ConsumerApplication {
 
   public static void main(String[] args) {
