+++ order/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
     implementation project(':message')      // message module 의존성 추가
-    implementation project(':domain')      // domain module 의존성 추가
+    implementation project(':common')      // common module 의존성 추가
 }
 
 bootJar.enabled = true
