+++ pay/src/main/java/com/commerce/saleday/pay/PayApplication.java
@@ -1,17 +1,18 @@
 package com.commerce.saleday.pay;
 
+import com.commerce.saleday.common.jpa.config.QuerydslConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
 @ComponentScan(basePackages = {
-    "com.commerce.saleday.domain",
+    "com.commerce.saleday.common",
     "com.commerce.saleday.order",
     "com.commerce.saleday.pay"
 })
+@Import(QuerydslConfig.class)// 같은 queryDslConfig쓰고 있음을 명시
 public class PayApplication {
 
   public static void main(String[] args) {
