+++ pay/src/main/java/com/commerce/saleday/pay/domain/model/Payment.java
@@ -1,6 +1,6 @@
 package com.commerce.saleday.pay.domain.model;
 
-import com.commerce.saleday.domain.common.BaseEntity;
+import com.commerce.saleday.common.model.BaseEntity;
 import com.commerce.saleday.order.domain.order.model.Orders;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
