+++ settings.gradle
@@ -2,6 +2,6 @@ rootProject.name = 'saleday'
 include 'consumer'
 include 'api'
 include 'message'
-include 'domain'
+include 'common'
 include 'pay'
 include 'order'
